modified validations
